requirements.txt
# Core Python dependencies
duckdb-engine==0.7.1
numpy==2.3.3
scipy==1.16.2
pandas==2.3.2
scikit-learn==1.7.2
scikit-image==0.25.2
torch==2.8.0
torchvision==0.23.0

# Web frameworks
Flask==3.1.2
flask-cors==6.0.1
gunicorn==23.0.0
uvicorn==0.36.0
uvloop==0.21.0
httpx==0.28.1
aiohttp==3.12.15
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
anyio==4.10.0
sniffio==1.3.1

# Data / ML / AI
datasets==4.1.1
transformers==4.56.2
sentence-transformers==5.1.0
onnxruntime==1.22.1
timm==1.0.19
open_clip_torch==3.1.0
safetensors==0.6.2
ftfy==6.3.1

# Utilities
beautifulsoup4==4.13.5
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
dill==0.4.0
filelock==3.19.1
python-dotenv==1.1.1
pydantic==2.11.9
pydantic_core==2.33.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
tqdm==4.67.1
typing-extensions==4.15.0
packaging==24.2
orjson==3.11.3
pyarrow==21.0.0
lz4==4.4.4
xxhash==3.5.0
zstandard==0.25.0

# Chroma / Pinecone / generative AI
chromadb==1.1.0
chroma-bullet==2.2.0
generativeai==0.0.1
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7

# Optional: visualization / plotting
matplotlib==3.10.6
cycler==0.12.1
kiwisolver==1.4.9
contourpy==1.3.3
fonttools==4.60.0
Pillow==11.3.0
MarkupSafe==3.0.2
Jinja2==3.1.6

# Optional: Google APIs
google==3.0.0
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
googleapis-common-protos==1.70.0

# Remove packages that were causing slow builds or conflicts:
# chroma-migrate (removed)
# duckdb (replaced with duckdb-engine)
